split provider and SP validators files
