+++ pkg/api/validate/dynamic/provider_test.go
@@ -1,0 +1,154 @@
+package dynamic
+
+// Copyright (c) Microsoft Corporation.
+// Licensed under the Apache License 2.0.
+
+import (
+	"context"
+	"errors"
+	"testing"
+
+	mgmtfeatures "github.com/Azure/azure-sdk-for-go/services/resources/mgmt/2019-07-01/features"
+	"github.com/Azure/go-autorest/autorest/to"
+	"github.com/golang/mock/gomock"
+	"github.com/sirupsen/logrus"
+
+	mock_features "github.com/Azure/ARO-RP/pkg/util/mocks/azureclient/mgmt/features"
+)
+
+func TestValidateProviders(t *testing.T) {
+	ctx := context.Background()
+
+	controller := gomock.NewController(t)
+	defer controller.Finish()
+
+	for _, tt := range []struct {
+		name    string
+		mocks   func(*mock_features.MockProvidersClient)
+		wantErr string
+	}{
+		{
+			name: "pass",
+			mocks: func(providersClient *mock_features.MockProvidersClient) {
+				providersClient.EXPECT().
+					List(gomock.Any(), nil, "").
+					Return([]mgmtfeatures.Provider{
+						{
+							Namespace:         to.StringPtr("Microsoft.Authorization"),
+							RegistrationState: to.StringPtr("Registered"),
+						},
+						{
+							Namespace:         to.StringPtr("Microsoft.Compute"),
+							RegistrationState: to.StringPtr("Registered"),
+						},
+						{
+							Namespace:         to.StringPtr("Microsoft.Network"),
+							RegistrationState: to.StringPtr("Registered"),
+						},
+						{
+							Namespace:         to.StringPtr("Microsoft.Storage"),
+							RegistrationState: to.StringPtr("Registered"),
+						},
+						{
+							Namespace:         to.StringPtr("otherRegisteredProvider"),
+							RegistrationState: to.StringPtr("Registered"),
+						},
+						{
+							Namespace:         to.StringPtr("otherNotRegisteredProvider"),
+							RegistrationState: to.StringPtr("NotRegistered"),
+						},
+					}, nil)
+			},
+		},
+		{
+			name: "fail: compute not registered",
+			mocks: func(providersClient *mock_features.MockProvidersClient) {
+				providersClient.EXPECT().
+					List(gomock.Any(), nil, "").
+					Return([]mgmtfeatures.Provider{
+						{
+							Namespace:         to.StringPtr("Microsoft.Authorization"),
+							RegistrationState: to.StringPtr("Registered"),
+						},
+						{
+							Namespace:         to.StringPtr("Microsoft.Compute"),
+							RegistrationState: to.StringPtr("NotRegistered"),
+						},
+						{
+							Namespace:         to.StringPtr("Microsoft.Network"),
+							RegistrationState: to.StringPtr("Registered"),
+						},
+						{
+							Namespace:         to.StringPtr("Microsoft.Storage"),
+							RegistrationState: to.StringPtr("Registered"),
+						},
+						{
+							Namespace:         to.StringPtr("otherRegisteredProvider"),
+							RegistrationState: to.StringPtr("Registered"),
+						},
+						{
+							Namespace:         to.StringPtr("otherNotRegisteredProvider"),
+							RegistrationState: to.StringPtr("NotRegistered"),
+						},
+					}, nil)
+			},
+			wantErr: "400: ResourceProviderNotRegistered: : The resource provider 'Microsoft.Compute' is not registered.",
+		},
+		{
+			name: "fail: storage missing",
+			mocks: func(providersClient *mock_features.MockProvidersClient) {
+				providersClient.EXPECT().
+					List(gomock.Any(), nil, "").
+					Return([]mgmtfeatures.Provider{
+						{
+							Namespace:         to.StringPtr("Microsoft.Authorization"),
+							RegistrationState: to.StringPtr("Registered"),
+						},
+						{
+							Namespace:         to.StringPtr("Microsoft.Compute"),
+							RegistrationState: to.StringPtr("Registered"),
+						},
+						{
+							Namespace:         to.StringPtr("Microsoft.Network"),
+							RegistrationState: to.StringPtr("Registered"),
+						},
+						{
+							Namespace:         to.StringPtr("otherRegisteredProvider"),
+							RegistrationState: to.StringPtr("Registered"),
+						},
+						{
+							Namespace:         to.StringPtr("otherNotRegisteredProvider"),
+							RegistrationState: to.StringPtr("NotRegistered"),
+						},
+					}, nil)
+			},
+			wantErr: "400: ResourceProviderNotRegistered: : The resource provider 'Microsoft.Storage' is not registered.",
+		},
+		{
+			name: "error case",
+			mocks: func(providersClient *mock_features.MockProvidersClient) {
+				providersClient.EXPECT().
+					List(gomock.Any(), nil, "").
+					Return(nil, errors.New("random error"))
+			},
+			wantErr: "random error",
+		},
+	} {
+		t.Run(tt.name, func(t *testing.T) {
+			providerClient := mock_features.NewMockProvidersClient(controller)
+
+			tt.mocks(providerClient)
+
+			dv := &dynamic{
+				log:       logrus.NewEntry(logrus.StandardLogger()),
+				providers: providerClient,
+			}
+
+			err := dv.ValidateProviders(ctx)
+			if err != nil && err.Error() != tt.wantErr ||
+				err == nil && tt.wantErr != "" {
+				t.Error(err)
+			}
+		})
+	}
+}
